requirements2.txt
anyio==3.6.2
asgiref==3.6.0
attrs==22.2.0
azure-core==1.26.3
azure-storage-blob==12.15.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==40.0.1
decorator==5.1.1
defusedxml==0.7.1
Django==4.1.7
django-allauth==0.54.0
django-cors-headers==3.14.0
django-phonenumber-field==7.0.2
django-rest-framework==0.1.0
django-rest-knox==4.2.0
django-restframework==0.0.1
djangorestframework==3.14.0
dynaconf==3.1.12
h11==0.14.0
httpcore==0.17.0
hvac==1.1.0
idna==3.4
isodate==0.6.1
Jinja2==3.1.2
knox==0.1.14
loguru==0.6.0
MarkupSafe==2.1.2
oauthlib==3.2.2
phonenumbers==8.13.8
psycopg2==2.9.5
pycparser==2.21
pyhcl==0.4.4
PyJWT==2.6.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
shortuuid==1.0.11
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.3
typing_extensions==4.5.0
tzdata==2023.2
urllib3==1.26.15
validators==0.20.0
win32-setctime==1.1.0
